Dockerfile, rework the command and termination
To avoid to wait 10s when we stop the spot container, this
patch performs some changes to only run the python binary in the container.
Tini is included in docker itself since 1.13 or greater, so it is removed
and the environments variables are handled in the application, that change avoids
to use a bash script to patch the configuration.