Update redis package and use pickle for serialize data
The new redis version is more strict about the kind of object
to store, we cannot use the same way that with the version 2.X.
In redis we want to store the SearchData class. Instead of
store field one by one we store the full serialized object with pickle.

Pickle is not very secure but we dont care for now because the redis
storage will be deploy in the same host that spot service. In a second time
we can use another way to serialize SearchData with msgpack and mashumaro
for example.